requirement.txt
openai-whisper
fairseq
praat-parselmouth
pyworld
faiss-cpu
torchcrepe

